splitting up radpattern functions into core / imaging
