+++ misc/docs/source/packages/obspy.core.event.rst
@@ -32,7 +32,6 @@
        header
        magnitude
        origin
-       radpattern
        source
 
     .. comment to end block
