+++ obspy/core/tests/test_event.py
@@ -32,6 +32,10 @@ class EventTestCase(unittest.TestCase):
     Test suite for obspy.core.event.Event
     """
     def setUp(self):
+        # directory where the test files are located
+        path = os.path.join(os.path.dirname(__file__), 'data')
+        self.path = path
+        self.image_dir = os.path.join(os.path.dirname(__file__), 'images')
         # Clear the Resource Identifier dict for the tests. NEVER do this
         # otherwise.
         ResourceIdentifier._ResourceIdentifier__resource_id_weak_dict.clear()
@@ -130,6 +134,16 @@ class EventTestCase(unittest.TestCase):
                       ev2.resource_id.get_referred_object())
         self.assertIs(ev.resource_id.get_referred_object(),
                       ev3.resource_id.get_referred_object())
+
+    def test_plot_farfield_without_quiver(self):
+        """
+        Tests to plot P/S wave farfield radiation pattern
+        """
+        ev = read_events("/path/to/CMTSOLUTION", format="CMTSOLUTION")[0]
+        with ImageComparison(self.image_dir, 'event_radiation_pattern.png') \
+                as ic:
+            ev.plot(kind=['s_sphere', 'p_sphere', 'beachball'],
+                    outfile=ic.name, show=False)
 
 
 class OriginTestCase(unittest.TestCase):
